# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Node",
    "description": "Project",
    "version": "0.1",
    "contact": {
      "email": "145s@ukr.net",
      "name": "Vlad",
      "url": "https://github.com/vladpanfilov1998"
    }
  },
  "paths": {
    "/auth/registration": {
      "post": {
        "description": "Registration",
        "operationId": "registration",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "firstName",
            "in": "body",
            "required": true,
            "description": "Minimum 3 and maximum 25 characters",
            "pattern": "Test"
          },
          {
            "type": "string",
            "name": "lastName",
            "in": "body",
            "required": true,
            "description": "Minimum 3 and maximum 25 characters",
            "pattern": "Test"
          },
          {
            "type": "number",
            "name": "age",
            "in": "body",
            "description": "Maximum age 120 years",
            "pattern": "22"
          },
          {
            "type": "string",
            "name": "phone",
            "in": "body",
            "required": true,
            "uniqueItems": true,
            "pattern": "0123456789",
            "description": "Phone length 10 characters"
          },
          {
            "$ref": "#/definitions/emailRequest"
          },
          {
            "$ref": "#/definitions/passwordRequest"
          },
          {
            "type": "file",
            "name": "avatar",
            "in": "formData",
            "description": "Only img format .gif, .jpeg, .pjpeg, .png, .svg+xml"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "accessToken": {
                  "$ref": "#/definitions/tokenResponse"
                },
                "refreshToken": {
                  "$ref": "#/definitions/tokenResponse"
                },
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "userEmail": {
                  "type": "string",
                  "example": "email@email.com"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/errorResponse"
            },
            "description": "Bad request"
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "description": "Refresh token pair",
        "operationId": "refresh",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "$ref": "#/definitions/tokenRequest",
            "description": "Refresh Token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/tokenPairResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
...
    "/posts/postId": {
      "patch": {
        "description": "Update Post",
        "operationId": "updatePost",
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "$ref": "#/definitions/tokenRequest",
            "description": "Access Token"
          },
          {
            "type": "string",
            "name": "title",
            "in": "body",
            "required": true,
            "pattern": "Test"
          },
          {
            "type": "string",
            "in": "body",
            "name": "text",
            "required": true,
            "pattern": "Test"
          }
        ],
        "responses": {
          "204": {
            "description": "Correct update"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/errorResponse"
            },
            "description": "Bad Request"
          },
          "401": {
            "schema": {
              "$ref": "#/definitions/errorResponse"
            },
            "description": "Unauthorized"
          },
          "404": {
            "schema": {
              "$ref": "#/definitions/errorResponse"
            },
            "description": "Not Found"
          }
        }
      }
    },
    "/users": {
      "get": {
        "description": "Get All Users",
        "operationId": "getUsers",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "page",
            "in": "query",
            "pattern": "1"
          },
          {
            "type": "string",
            "name": "perPage",
            "in": "query",
            "pattern": "1"
          },
          {
            "type": "string",
            "name": "firstName",
            "in": "query",
            "pattern": "Test"
          },
          {
            "type": "string",
            "name": "lastName",
            "in": "query",
            "pattern": "Test"
          },
          {
            "type": "number",
            "name": "age",
            "in": "query",
            "pattern": "22"
          },
          {
            "type": "string",
            "name": "phone",
            "in": "query",
            "pattern": "0123456789"
          },
          {
            "type": "string",
            "name": "email",
            "in": "query",
            "pattern": "mail@mail.com"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "page": {
                  "type": "number",
                  "example": 1
                },
                "perPage": {
                  "type": "number",
                  "example": 10
                },
                "totalCount": {
                  "type": "number",
                  "example": 100
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/userResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/:id": {
      "delete": {
        "description": "Delete User",
        "operationId": "deleteUser",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "description": "Access Token",
            "$ref": "#/definitions/tokenRequest"
          },
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "pattern": "1",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Correct Delete"
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "tokenResponse": {
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjE1LCJ1c2VyRW1haWwiOiJ6eNGE0YbRg2N2QG1haWwuY29tIiwiaWF0IjoxNjUwMDM5MDkyLCJlsdgdhgdhfdhAwMzk5OTJ9.VQBNSWAGyqqkWxR3tyO8HeE4dPLYYTrOluqvzftNjWk",
      "description": "Token"
    },
    "errorResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string",
          "example": "Something went wrong",
          "description": "Standard error response"
        }
      }
    },
...
    "userResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1,
          "description": "Id"
        },
        "firstName": {
          "type": "string",
          "example": "Test",
          "description": "First Name"
        },
        "lastName": {
          "type": "string",
          "example": "Test",
          "description": "Last Name"
        },
        "age": {
          "type": "number",
          "example": 1,
          "description": "Age"
        },
        "phone": {
          "type": "string",
          "example": "0123456789",
          "description": "Phone"
        },
        "email": {
          "type": "string",
          "example": "test@test.com",
          "description": "Email"
        },
        "createdAt": {
          "type": "string",
          "example": "2022-04-15T13:11:30.000Z",
          "description": "createdAt"
        },
        "deletedAt": {
          "type": [
            "string",
            "null"
          ],
          "example": [
            "2022-04-15T13:11:30.000Z",
            null
          ],
          "description": "deletedAt"
        }
      }
    },
    "tokenPairResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "$ref": "#/definitions/tokenResponse"
        },
        "refreshToken": {
          "$ref": "#/definitions/tokenResponse"
        },
        "user": {
          "$ref": "#/definitions/userResponse"
        }
      }
    },
...
    "tokenRequest": {
      "type": "string",
      "name": "Authorization",
      "in": "header",
      "required": true,
      "pattern": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjE1LCJ1c2VyRW1haWwiOiJ6eNGE0YbRg2N2QG1haWwuY29tIiwiaWF0IjoxNjUwMDM5MDkyLCJlsdgdhgdhfdhAwMzk5OTJ9.VQBNSWAGyqqkWxR3tyO8HeE4dPLYYTrOluqvzftNjWk"
    },
    "emailRequest": {
      "type": "string",
      "name": "email",
      "in": "body",
      "required": true,
      "uniqueItems": true,
      "pattern": "email@email.com"
    },
    "passwordRequest": {
      "type": "string",
      "name": "password",
      "in": "body",
      "required": true,
      "pattern": "Test^^^12345",
      "description": "Minimum eight and maximum 25 characters, at least one uppercase letter, one lowercase letter, one number and one special character"
    }
  }
}